pyproject.toml
[tool.poetry]
name = "webview-py-editor"
version = "0.1.0"
description = "Demo Python Webview code editor"
authors = [":Barry-Thomas-Paul: Moss <4193389+Amourspirit@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
pywebview = [
	{extras = ["qt"], version = "^5.3.2", markers = "sys_platform != 'win32'"},
	{extras = ["cef"], version = "^5.3.2", markers = "sys_platform == 'win32'"}
]
jedi = "^0.19.1"
typing-extensions = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
